+++ internal/fleet/agentpolicy/testdata/TestAccResourceAgentPolicyPolicyIDValidation/traversal_policy_id/main.tf
@@ -1,0 +1,12 @@
+provider "elasticstack" {
+  elasticsearch {}
+  kibana {}
+}
+
+resource "elasticstack_fleet_agent_policy" "test_policy" {
+  name            = "unit-test-policy"
+  namespace       = "default"
+  policy_id       = "my..policy"
+  monitor_logs    = true
+  monitor_metrics = false
+}